Added MAC adress to topics

+++ esp32_projects/RBG_Action/src/HTTPClient.cpp
@@ -15,7 +15,7 @@ void updateFirmware(byte *pay, unsigned int length)
 
   HTTPClient http;
   http.begin(firmwareURL);
-  int httpCode = http.GET();
+  uint16_t httpCode = http.GET();
 
   if (httpCode > 0)
   {
